Events
Now you can execute custom process for each actions.

+++ Controller/ManageController.php
@@ -1,15 +1,31 @@
 <?php
 
+/*
+ * This file is part of the DBManagerBundle package.
+ *
+ * (c) FOUQUET <https://github.com/hugo082/DBManagerBundle>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @author Hugo Fouquet <hugo.fouquet@epita.fr>
+ */
+
 namespace DB\ManagerBundle\Controller;
 
+use DB\ManagerBundle\DBManagerEvents;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use DB\ManagerBundle\DependencyInjection\Configuration;
+use DB\ManagerBundle\Event\ActionEvent;
+use DB\ManagerBundle\Checker\EntityManager as Checker;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface as Dispatcher;
 
 class ManageController extends Controller
 {
     public function indexAction()
     {
+        /** @var $checker Checker */
         $checker = $this->get('db.manager.checker');
         $entities = $checker->getEntities();
 
@@ -20,6 +36,9 @@ class ManageController extends Controller
 
     public function listAction(Request $request, $name)
     {
+        /** @var $dispatcher Dispatcher */
+        $dispatcher = $this->get('event_dispatcher');
+        /** @var $checker Checker */
         $checker = $this->get('db.manager.checker');
         $eInfo = $checker->getEntity($name, Configuration::PERM_LIST);
 
@@ -31,11 +50,7 @@ class ManageController extends Controller
             $form = $this->createForm($eInfo['fullFormType'], $e);
             $form->handleRequest($request);
             if ($form->isSubmitted() && $form->isValid()) {
-                $em = $this->getDoctrine()->getManager();
-                $em->persist($e);
-                $em->flush();
-                $this->addFlash('success','Your entity have been added');
-                return $this->redirectToRoute('db.manager.list', array('name' => $name));
+                return $this->addFormReception($eInfo, $e, $dispatcher, $name);
             }
             $form = $form->createView();
         }
@@ -45,13 +60,15 @@ class ManageController extends Controller
             'eInfo' => $eInfo,
             'all' => $all,
             'form' => $form,
-            'action' => array( 'name' => 'list', 'formType' => 'add'),
-            'settings' => $checker->getSettings()
+            'action' => array( 'name' => Configuration::PERM_ADD, 'formType' => Configuration::PERM_ADD)
         ));
     }
 
     public function addAction(Request $request, $name)
     {
+        /** @var $dispatcher Dispatcher */
+        $dispatcher = $this->get('event_dispatcher');
+        /** @var $checker Checker */
         $checker = $this->get('db.manager.checker');
         $eInfo = $checker->getEntity($name, Configuration::PERM_ADD);
         $checker->checkObjectPermission($eInfo, NULL, Configuration::PERM_ADD);
@@ -60,11 +77,7 @@ class ManageController extends Controller
         $form = $this->createForm($eInfo['fullFormType'], $e);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($e);
-            $em->flush();
-            $this->addFlash('success','Your entity have been added');
-            return $this->redirectToRoute('db.manager.list', array('name' => $name));
+            return $this->addFormReception($eInfo, $e, $dispatcher, $name);
         }
 
         return $this->render($eInfo['mainView'], array(
@@ -72,13 +85,15 @@ class ManageController extends Controller
             'eInfo' => $eInfo,
             'all' => NULL,
             'form' => $form->createView(),
-            'action' => array( 'name' => 'add', 'formType' => 'add'),
-            'settings' => $checker->getSettings()
+            'action' => array( 'name' => Configuration::PERM_ADD, 'formType' => Configuration::PERM_ADD)
         ));
     }
 
     public function editAction(Request $request, $name, $id)
     {
+        /** @var $dispatcher Dispatcher */
+        $dispatcher = $this->get('event_dispatcher');
+        /** @var $checker Checker */
         $checker = $this->get('db.manager.checker');
         $eInfo = $checker->getEntity($name, Configuration::PERM_EDIT);
 
@@ -91,10 +106,15 @@ class ManageController extends Controller
         $form = $this->createForm($eInfo['fullFormType'], $e);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($e);
-            $em->flush();
-            $this->addFlash('success','Your entity have updated');
+            $event = new ActionEvent($eInfo, $e, array('success', 'Your entity have been updated'));
+            $dispatcher->dispatch(DBManagerEvents::ACTION_EDIT_BEFORE, $event);
+
+            if (!$event->isExecuted()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->persist($e);
+                $em->flush();
+            }
+            $this->addFlash($event->getFlashTitle(),$event->getFlashMessage());
             return $this->redirectToRoute('db.manager.list', array('name' => $name));
         }
         return $this->render($eInfo['mainView'], array(
@@ -102,24 +122,53 @@ class ManageController extends Controller
             'eInfo' => $eInfo,
             'all' => $all,
             'form' => $form->createView(),
-            'action' => array( 'name' => 'edit', 'formType' => 'edit'),
-            'settings' => $checker->getSettings()
+            'action' => array( 'name' => Configuration::PERM_EDIT, 'formType' => Configuration::PERM_EDIT)
         ));
     }
 
     public function removeAction($name, $id)
     {
+        /** @var $dispatcher Dispatcher */
+        $dispatcher = $this->get('event_dispatcher');
+        /** @var $checker Checker */
         $checker = $this->get('db.manager.checker');
         $eInfo = $checker->getEntity($name, Configuration::PERM_REMOVE);
 
         $e = $checker->getEntityObject($eInfo, Configuration::PERM_REMOVE, $id);
         if ($e) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($e);
-            $em->flush();
-            $this->addFlash('success','Your entity have been removed');
+            $event = new ActionEvent($eInfo, $e, array('success', 'Your entity have been removed'));
+            $dispatcher->dispatch(DBManagerEvents::ACTION_REMOVE_BEFORE, $event);
+
+            if (!$event->isExecuted()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($e);
+                $em->flush();
+            }
+            $this->addFlash($event->getFlashTitle(),$event->getFlashMessage());
         } else
             $this->addFlash('danger','Your entity could not be deleted');
+        return $this->redirectToRoute('db.manager.list', array('name' => $name));
+    }
+
+    /**
+     * Process on add form reception
+     *
+     * @param array $eInfo
+     * @param $e
+     * @param Dispatcher $dispatcher
+     * @param string $name
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     */
+    private function addFormReception(array $eInfo, $e, Dispatcher $dispatcher, string $name){
+        $event = new ActionEvent($eInfo, $e, array('success', 'Your entity have been added'));
+        $dispatcher->dispatch(DBManagerEvents::ACTION_ADD_BEFORE, $event);
+
+        if (!$event->isExecuted()) {
+            $em = $this->getDoctrine()->getManager();
+            $em->persist($e);
+            $em->flush();
+        }
+        $this->addFlash($event->getFlashTitle(),$event->getFlashMessage());
         return $this->redirectToRoute('db.manager.list', array('name' => $name));
     }
 }
